+++ tests/test_checker/test_noqa.py
@@ -125,6 +125,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS349': 1,
     'WPS350': 1,
     'WPS351': 1,
+    'WPS352': 1,
 
     'WPS400': 0,
     'WPS401': 0,